FIX: Bug 1032969
Added a button that calculates the progress
of selected runs. This applies to places including
plan page, run search page, and dvanced run search.
And, noticably but badly, this commit also includes
one other change that applies shift-selection for
run list in those places too. A bugfix is made to
the shift-selection plugin.