REV: reduce the number of SQLs from more than 200 to 11
This commit introduces a new class as the iteration proxy that iterates objects
and its associated data only at once. In this way, not only saves many SQL
queries, but also reduce the memory usage. For more details, such as how to use
it, refer to `tcms.core.utils.QuerySetIterationProxy'. Oh, I know I haven't
found a better name for it.

Change-Id: I8f18229439af4c6bca6b09e92256a6ad1c2e069f