REV: optimize assign cases page loading and update
TCMS-232 - [Performance] speed up loading /run/${run_id}/assigncase and update

Enlarge the scope of the performance improvement. Not only optimize the page
loading, but also the assignment process. Because the original implementation
of View method mixes the two process together. It is not possible to make
changes to page loading only.

Change-Id: I5a95fb0322807383c5f1499b1eeff87eba918912